Print contents of source file in error messages
At this point, error messages contain location information of the error.
In practice, this means that the the user must locate the error by
searching the source file for the line and in some cases scrolling to
the correct location. This process can be facilitated if the errors
printed code from the source file and underlined the token mentioned
in the `location` field. Print this information in all errors.